+++ xeHentai/config.py
@@ -31,6 +31,8 @@ rpc_interface = 'localhost'
 rpc_port = None
 # jsonrpc secret string
 rpc_secret = None
+# auto open browser on rpc start
+rpc_open_browser = True
 
 # make an archive (.zip) after download and delete directory
 make_archive = False
@@ -63,3 +65,5 @@ save_tasks = False
 # delete files when deleting a task
 delete_task_files = False
 
+# retry a connection if per thread speed is lower than this value, unit is KB per second
+low_speed_threshold = 10