+++ xeHentai/rpc.py
@@ -10,6 +10,9 @@ import zipfile
 import traceback
 from hashlib import md5
 from threading import Thread
+import zlib
+import requests
+import pickle
 from .const import *
 from .const import __version__
 from .i18n import i18n
@@ -26,12 +29,15 @@ else:
     from urlparse import urlparse
 
 cmdre = re.compile("([a-z])([A-Z])")
-pathre = re.compile("/(?:jsonrpc|img|zip)")
-imgpathre = re.compile("/img")
-zippathre = re.compile("/zip")
+pathre = re.compile("/(?:jsonrpc|img/|zip/|static/|ui/$)")
+staticre = re.compile("/static/")
+imgpathre = re.compile("/img/")
+zippathre = re.compile("/zip/")
+
+version_str = "xeHentai/%s" % __version__
 
 class RPCServer(Thread):
-    def __init__(self, xeH, bind_addr, secret = None, logger = None, exit_check = None):
+    def __init__(self, xeH, bind_addr, secret = None, open_browser = True, logger = None, exit_check = None):
         Thread.__init__(self, name = "rpc")
         Thread.setDaemon(self, True)
         self.xeH = xeH
@@ -39,6 +45,7 @@ class RPCServer(Thread):
         self.secret = secret
         self.logger = logger
         self.server = None
+        self.open_browser = open_browser
         self._exit = exit_check if exit_check else lambda x:False
 
     def run(self):
@@ -48,13 +55,22 @@ class RPCServer(Thread):
             self.logger.error(i18n.RPC_CANNOT_BIND % traceback.format_exc())
         else:
             self.logger.info(i18n.RPC_STARTED % (self.bind_addr[0], self.bind_addr[1]))
+            url = "http://%s:%s/ui/#host=%s,port=%s,https=no" % (
+                self.bind_addr[0], self.bind_addr[1],
+                self.bind_addr[0], self.bind_addr[1]
+            )
+            if self.secret:
+                url = url + ",token=" + self.secret
+            if self.open_browser:
+                import webbrowser
+                webbrowser.open(url)
+            else:
+                self.logger.info(i18n.RPC_WEBUI_PATH % url)
             while not self._exit("rpc"):
                 self.server.handle_request()
 
-def is_file_obj(obj):
-    if PY3K:
-        return isinstance(obj, IOBase)
-    return isinstance(obj, file)
+def is_readable_obj(obj):
+    return hasattr(obj, "read")
 
 def is_str_obj(obj):
     if PY3K:
@@ -78,7 +94,7 @@ def gen_thumbnail(fh, args):
         return fh, False
     size = (int(args['w']) if 'w' in args else int(args['h']),
             int(args['h']) if 'h' in args else int(args['w']))
-    if not is_file_obj(fh):
+    if not is_readable_obj(fh):
         fh = StringIO(fh)
     with Image.open(fh) as img:
         img.thumbnail(size)
@@ -114,18 +130,42 @@ def path_filter(func):
         func(self)
     return f
 
+def load_cache():
+    if os.path.exists(STATIC_CACHE_FILE):
+        try:
+            with open(STATIC_CACHE_FILE, "rb") as f:
+                r = zlib.decompress(f.read())
+                r = pickle.loads(r)
+                if 'v' in r or r['v'] == STATIC_CACHE_VERSION:
+                    return r
+        except:
+            pass
+    return { "v": STATIC_CACHE_VERSION }
+
+def save_cache(static_cache):
+    r = pickle.dumps(static_cache)
+    r = zlib.compress(r)
+    with open(STATIC_CACHE_FILE, "wb") as f:
+        f.write(r)
+
+static_cache = load_cache()
 class Handler(BaseHTTPRequestHandler):
 
     def __init__(self, xeH, secret, *args):
         self.secret = secret
         self.args = args
         self.xeH = xeHentaiRPCExtended(xeH, secret)
+        self.http = requests.Session()
         BaseHTTPRequestHandler.__init__(self, *args)
 
     def version_string(self):
-        return "xeHentai/%s" % __version__
+        return version_str
     
     def serve_file(self, f):
+        if hasattr(self.xeH, "_monitor"):
+            _task = self.xeH._monitor.task
+            # needed to lock between archiver
+            _task._f_lock.acquire()
         f.seek(0, os.SEEK_END)
         size = f.tell()
         self.xeH.logger.verbose("GET %s 200 %d %s" % (self.path, size, self.client_address[0]))
@@ -137,6 +177,8 @@ class Handler(BaseHTTPRequestHandler):
             if not buf:
                 break
             self.wfile.write(buf)
+        if hasattr(self.xeH, "_monitor"):
+            _task._f_lock.release()
         return size
 
     def do_OPTIONS(self):
@@ -153,10 +195,11 @@ class Handler(BaseHTTPRequestHandler):
         code = 200
         rt = b''
         mime = "text/html"
+        path = self.path
         while True:
-            if imgpathre.match(self.path):
-                args = dict(q.split("=") for q in urlparse(self.path).query.split("&") if q)
-                _ = urlparse(self.path).path.split("/")
+            if imgpathre.match(path):
+                args = dict(q.split("=") for q in urlparse(path).query.split("&") if q)
+                _ = urlparse(path).path.split("/")
                 if len(_) < 5:
                     code = 400
                     break
@@ -187,9 +230,9 @@ class Handler(BaseHTTPRequestHandler):
                 rt, _error = gen_thumbnail(rt, args)
                 if _error:
                     self.xeH.logger.warning("RPC: PIL needed for generating thumbnail")
-            elif zippathre.match(self.path):
+            elif zippathre.match(path):
                 # args = urlparse(_).query
-                _ = urlparse(self.path).path.split("/")
+                _ = urlparse(path).path.split("/")
                 if len(_) < 5:
                     code = 400
                     break
@@ -207,6 +250,52 @@ class Handler(BaseHTTPRequestHandler):
                     code = 404
                     break
                 rt = open(f, 'rb')
+            elif path == "/ui/" or staticre.match(path):
+                if path == "/ui/":
+                    path = "/"
+                while True:
+                    cache_rt = None
+                    should_clear_cache = False
+                    headers = { "User-Agent":  version_str }
+                    if path in static_cache:
+                        cache_rt, mime, tm, lms = static_cache[path]
+                        if PY3K and not isinstance(cache_rt, bytes):
+                            cache_rt = bytes(cache_rt, 'ascii')
+                        if time.time() - STATIC_CACHE_TTL < tm:
+                            rt = StringIO(cache_rt)
+                            break
+                        should_clear_cache = True
+                        headers['If-Modified-Since'] = lms
+
+                    req_start_tm = time.time()
+                    r = None
+                    try:
+                        r = self.http.get("http://xehentai.yooooo.us%s?_=%d" %(path, time.time()), headers=headers)
+                    except Exception as ex:
+                        self.xeH.logger.warn("error pulling %s from remote server: %s" % (path, err))
+                    self.xeH.logger.verbose("%.2fs taken to pull %s from remote server %s bytes" % (
+                                            time.time() - req_start_tm, path, r and len(r.content) or 0))
+                    if r and r.status_code == 200:
+                        rt = StringIO(r.content)
+                        mime = r.headers['Content-type']
+                        if should_clear_cache:
+                            # clear all keys, since the js/css hash may change
+                            static_cache.clear()
+                        static_cache[path] = [r.content, mime, time.time(), r.headers['Last-Modified']]
+                        save_cache(static_cache)
+                    elif r and r.status_code == 304:
+                        # so this is tricky: if we hit /ui/ first and it's not expired
+                        # then all other assets should not expire
+                        if path == "/":
+                            for k in static_cache:
+                                static_cache[k][2] = time.time()
+                        rt = StringIO(cache_rt)
+                    elif cache_rt:
+                        self.xeH.logger.warn("serving stale cache %s" % (path))
+                        rt = StringIO(cache_rt)
+                    else:
+                        rt = jsonrpc_resp({"id":None}, error_code = ERR_RPC_INVALID_REQUEST)
+                    break
             else:
                 # fallback to rpc request
                 rt = jsonrpc_resp({"id":None}, error_code = ERR_RPC_INVALID_REQUEST)
@@ -217,7 +306,7 @@ class Handler(BaseHTTPRequestHandler):
         self.send_header("Access-Control-Allow-Origin", "*")
         self.send_header("Content-Type", mime)
         
-        if is_file_obj(rt):
+        if is_readable_obj(rt):
             size = self.serve_file(rt)
             rt.close()
         else:
@@ -279,6 +368,9 @@ class Handler(BaseHTTPRequestHandler):
                     break
             self.xeH.logger.verbose("RPC from: %s, cmd: %s, params: %s" % (self.client_address[0], cmd, params))
             try:
+                # pop out token if extra token is found
+                if len(params[0]) > 0 and 'token:' in params[0][0]:
+                    del params[0][0]
                 cmd_rt = getattr(self.xeH, cmd_r)(*params[0], **params[1])
             except (ValueError, TypeError) as ex:
                 self.xeH.logger.verbose("RPC exec error:\n%s" % traceback.format_exc())
@@ -315,7 +407,8 @@ class xeHentaiRPCExtended(object):
     def get_info(self):
         ret = {"version": self.verstr,
             "threads_zombie": 0, "threads_running": 0,
-            "queue_pending": 0, "queue_finished": 0
+            "queue_pending": 0, "queue_finished": 0,
+            "download_speed": 0,
         }
         if hasattr(self, '_monitor'):
             ret['threads_running'] = len(self._monitor.thread_last_seen)
@@ -323,6 +416,7 @@ class xeHentaiRPCExtended(object):
             if self._monitor.task.state > TASK_STATE_PAUSED and self._monitor.task.img_q:
                 ret['queue_pending'] = self._monitor.task.img_q.qsize()
                 ret['queue_finished'] = self._monitor.task.meta['finished']
+                ret['download_speed'] = self._monitor.download_speed
             else:
                 ret['queue_pending'] = 0
                 ret['queue_finished'] = 0
@@ -348,7 +442,7 @@ class xeHentaiRPCExtended(object):
             return ERR_TASK_LEVEL_UNDEF, None
         lv = globals()[level]
         rt = [{_k:_v for _k, _v in v.to_dict().items() if _k not in
-            ('reload_map', 'filehash_map', 'renamed_map', 'img_q', 'page_q')}
+            ('reload_map', 'duplicate_map', 'renamed_map', 'logger', 'img_q', 'page_q')}
                  for _, v in self._all_tasks.items() if 
                     (reverse_mode and v.state != lv) or (not reverse_mode and v.state == lv)]
         return ERR_NO_ERROR, rt