+++ xeHentai/worker.py
@@ -14,7 +14,7 @@ from threading import Thread, RLock
 from . import util
 from .const import *
 from .i18n import i18n
-from .proxy import PoolException
+from .proxy import PoolException, LowSpeedException
 if PY3K:
     from queue import Queue, Empty
     from urllib.parse import urlparse, urlunparse
@@ -137,11 +137,42 @@ class HttpReq(object):
             retry += 1
         return _filter(_FakeResponse(url_history[0]), suc, fail)
 
+# speed statistics with ring buffer
+class speed_checker(object):
+    def __init__(self, cnt=5):
+        self.cnt = cnt
+        self.speed_buffer = []
+        self.reset()
 
+    def check(self, l):
+        self.current_bytes += l
+        self.current_tm = time.time()
+        if self.current_tm - self.last_tm > 1:
+            self.speed_buffer.append((self.current_bytes-self.last_bytes)/(self.current_tm-self.last_tm))
+            while len(self.speed_buffer) > self.cnt:
+                self.speed_buffer.pop(0)
+            self.last_tm = self.current_tm
+            self.last_bytes = self.current_bytes
+        return
+
+    def calc(self, full=False):
+        if len(self.speed_buffer) == 0:
+            return 0
+        elif full and len(self.speed_buffer) < self.cnt:
+            return 0
+        return sum(self.speed_buffer)/len(self.speed_buffer)
+
+    def reset(self):
+        self.last_tm = time.time()
+        self.last_bytes = 0
+        self.current_bytes = 0
+        self.current_tm = 0
+        if self.speed_buffer:
+            self.speed_buffer = []
 
 class HttpWorker(Thread, HttpReq):
     def __init__(self, tname, task_queue, flt, suc, fail, headers={}, proxy=None, proxy_policy=None,
-            retry=3, timeout=10, logger=None, keep_alive=None, stream_mode=False):
+            retry=3, timeout=10, logger=None, keep_alive=None, stream_mode=False, lowspeed_threshold=None):
         """
         Construct a new 'HttpWorker' obkect
 
@@ -171,6 +202,8 @@ class HttpWorker(Thread, HttpReq):
         self.f_suc = suc
         self.f_fail = fail
         self.stream_mode = stream_mode
+        self.stream_speed = None
+        self.lowspeed_threshold = lowspeed_threshold
         # if we don't checkin in this zombie_threshold time, monitor will regard us as zombie
         self.zombie_threshold = timeout * (retry + 1) 
         self.run_once = False
@@ -183,11 +216,22 @@ class HttpWorker(Thread, HttpReq):
         self.logger.verbose("t-%s start" % self.name)
         _stream_cb = None
         if self.stream_mode:
-            _stream_cb = lambda x:self._keepalive(self)
+            self.stream_speed = speed_checker()
+            def f(d):
+                self.stream_speed.check(len(d))
+                if self.lowspeed_threshold:
+                    speed = self.stream_speed.calc(full=True)
+                    if 0 < speed < self.lowspeed_threshold:
+                        raise LowSpeedException("")
+                self._keepalive(self)
+            _stream_cb = f
         while not self._keepalive(self) and not self._exit(self):
             try:
                 url = self.task_queue.get(False)
             except Empty:
+                # set back to 0 when waiting
+                if self.stream_speed:
+                    self.stream_speed.reset()
                 time.sleep(1)
                 continue
             self.run_once = True
@@ -196,6 +240,13 @@ class HttpWorker(Thread, HttpReq):
             except PoolException as ex:
                 self.logger.warning("%s-%s %s" % (i18n.THREAD, self.tname, str(ex)))
                 break
+            except LowSpeedException as ex:
+                self.logger.warning(i18n.THREAD_SPEED_TOO_LOW % (
+                    self.tname,
+                    util.human_size(self.stream_speed.calc(full=True)),
+                    util.human_size(self.lowspeed_threshold),
+                ))
+                self.flt(_FakeResponse(url), self.f_suc, self.f_fail)
             except Exception as ex:
                 self.logger.warning(i18n.THREAD_UNCAUGHT_EXCEPTION % (self.tname, traceback.format_exc()))
                 self.flt(_FakeResponse(url), self.f_suc, self.f_fail)
@@ -251,6 +302,7 @@ class Monitor(Thread):
         self.task = task
         self._exit = exit_check if exit_check else lambda x: False
         self._cleaning_up = False
+        self.download_speed = 0
         if os.name == "nt":
             self.set_title = lambda s:os.system("TITLE %s" % (
                 s if PY3K else s.encode(CODEPAGE, 'replace')))
@@ -314,12 +366,12 @@ class Monitor(Thread):
     #     print(self.task.list_q.qsize())
 
     def _check_vote(self):
-        if False and ERR_IMAGE_RESAMPLED in self.vote_result and ERR_IMAGE_RESAMPLED not in self.vote_cleared:
-            self.logger.warning(i18n.TASK_START_PAGE_RESCAN % self.task.guid)
-            self._rescan_pages()
-            self.task.meta['has_ori'] = True
-            self.vote_cleared.add(ERR_IMAGE_RESAMPLED)
-        elif ERR_QUOTA_EXCEEDED in self.vote_result and \
+        # if False and ERR_IMAGE_RESAMPLED in self.vote_result and ERR_IMAGE_RESAMPLED not in self.vote_cleared:
+        #     self.logger.warning(i18n.TASK_START_PAGE_RESCAN % self.task.guid)
+        #     self._rescan_pages()
+        #     self.task.meta['has_ori'] = True
+        #     self.vote_cleared.add(ERR_IMAGE_RESAMPLED)
+        if ERR_QUOTA_EXCEEDED in self.vote_result and \
             ERR_QUOTA_EXCEEDED not in self.vote_cleared and \
             self.vote_result[ERR_QUOTA_EXCEEDED] >= len(self.thread_last_seen):
             self.logger.error(i18n.TASK_STOP_QUOTA_EXCEEDED % self.task.guid)
@@ -335,6 +387,7 @@ class Monitor(Thread):
         while len(self.thread_last_seen) > 0:
             intv += 1
             self._check_vote()
+            total_speed = 0
             for k in list(self.thread_last_seen.keys()):
                 _zombie_threshold = self.thread_ref[k].zombie_threshold if k in self.thread_ref else 30
                 if time.time() - self.thread_last_seen[k] > _zombie_threshold:
@@ -344,13 +397,18 @@ class Monitor(Thread):
                     else:
                         self.logger.warning(i18n.THREAD_SWEEP_OUT % k)
                     del self.thread_last_seen[k]
+                # if thread is not a zombie, add to speed sum
+                elif k in self.thread_ref and self.thread_ref[k].stream_speed:
+                    total_speed += self.thread_ref[k].stream_speed.calc()
+            self.download_speed = total_speed
             if intv == CHECK_INTERVAL:
-                _ = "%s %dR/%dZ, %s %dR/%dD" % (
+                _ = "%s %dR/%dZ, %s %dR/%dD, %s/s" % (
                     i18n.THREAD,
                     len(self.thread_last_seen), len(self.thread_zombie),
                     i18n.QUEUE,
                     self.task.img_q.qsize(),
-                    self.task.meta['finished'])
+                    self.task.meta['finished'],
+                    util.human_size(total_speed))
                 self.logger.info(_)
                 self.set_title(_)
                 intv = 0
@@ -358,8 +416,13 @@ class Monitor(Thread):
                 if last_finished != self.task.meta['finished']:
                     last_change = time.time()
                     last_finished = self.task.meta['finished']
-                else:
-                    if time.time() - last_change > STUCK_INTERVAL:
+                elif time.time() - last_change > STUCK_INTERVAL:
+                    self.logger.info(i18n.TASK_UNFINISHED % (self.task.guid, self.task.get_fid_unfinished()))
+                    if total_speed > 0:
+                        # reset last_change
+                        last_change = time.time()
+                        self.logger.warning(i18n.TASK_SLOW % self.task.guid)
+                    else:
                         self.logger.warning(i18n.TASK_STUCK % self.task.guid)
                         break
             time.sleep(0.5)